Moving plotting function into species_plotting out of html_tools.py
